Extract position calculation code
