+++ verifier/src/main/java/com/aws/jverify/verifier/PositionCalculator.java
@@ -1,0 +1,135 @@
+package com.aws.jverify.verifier;
+
+import com.aws.jverify.generated.Token;
+import com.sun.tools.javac.tree.JCTree;
+import com.sun.tools.javac.tree.TreeInfo;
+import com.sun.tools.javac.util.Position;
+
+import java.io.IOException;
+
+public class PositionCalculator {
+    JCTree.JCCompilationUnit compilationUnit;
+
+    public PositionCalculator(JCTree.JCCompilationUnit compilationUnit) {
+        this.compilationUnit = compilationUnit;
+    }
+
+    public int getStartPos(JCTree tree) {
+        return switch(tree) {
+            case JCTree.JCClassDecl classDecl -> getClassNamePosition(classDecl);
+            case JCTree.JCMethodDecl methodDecl -> getMethodNamePosition(methodDecl);
+            default -> tree.getStartPosition();
+        };
+    }
+
+    private int getMethodNamePosition(JCTree.JCMethodDecl methodDecl) {
+        CharSequence source;
+        try {
+            source = compilationUnit.getSourceFile().getCharContent(true);
+        } catch (IOException e) {
+            return Position.NOPOS;
+        }
+        String sourceText = source.toString();
+
+        int pos = methodDecl.pos;
+
+        if (methodDecl.mods != null && methodDecl.mods.pos != Position.NOPOS) {
+            // not sure if this statement is ever useful
+            pos = Math.max(pos, getEndPos(methodDecl.mods));
+        }
+
+        String methodName = methodDecl.name.toString();
+        boolean isConstructor = TreeInfo.isConstructor(methodDecl);
+
+        if (isConstructor) {
+            if (methodDecl.typarams != null && !methodDecl.typarams.isEmpty()) {
+                JCTree.JCTypeParameter last = methodDecl.typarams.last();
+                pos = getEndPos(last);
+            }
+            return pos;
+        } else {
+            if (methodDecl.typarams != null && !methodDecl.typarams.isEmpty()) {
+                pos = getEndPos(methodDecl.typarams.last());
+            }
+
+            if (methodDecl.restype != null) {
+                pos = getEndPos(methodDecl.restype);
+            }
+        }
+
+        while (pos < sourceText.length()) {
+            while (pos < sourceText.length() && Character.isWhitespace(sourceText.charAt(pos))) {
+                pos++;
+            }
+
+            if (pos + methodName.length() <= sourceText.length() &&
+                    sourceText.regionMatches(pos, methodName, 0, methodName.length())) {
+                // Verify this is the actual method name by checking if it's followed by a '('
+                int nextCharPos = pos + methodName.length();
+                while (nextCharPos < sourceText.length() && Character.isWhitespace(sourceText.charAt(nextCharPos))) {
+                    nextCharPos++;
+                }
+                if (nextCharPos < sourceText.length() && sourceText.charAt(nextCharPos) == '(') {
+                    return pos;
+                }
+            }
+
+            pos++;
+        }
+
+        return Position.NOPOS;
+    }
+
+    public int getEndPos(JCTree node) {
+        return TreeInfo.getEndPos(node, compilationUnit.endPositions);
+    }
+
+    private int getClassNamePosition(JCTree.JCClassDecl classDecl) {
+        CharSequence source;
+        try {
+            source = this.compilationUnit.getSourceFile().getCharContent(true);
+        } catch (IOException e) {
+            return Position.NOPOS;
+        }
+
+        int pos = classDecl.pos;
+
+        if (classDecl.mods != null && classDecl.mods.pos != Position.NOPOS) {
+            // Not sure if this statement is useful
+            pos = Math.max(pos, getEndPos(classDecl.mods));
+        }
+
+        // Find the class/interface/enum/record keyword and skip it
+        String sourceText = source.toString();
+        String[] keywords = {"class", "interface", "enum", "record"};
+
+        for (String keyword : keywords) {
+            int keywordPos = sourceText.indexOf(keyword + " ", pos);
+            if (keywordPos >= pos) {
+                // Found the keyword, the class name starts after it and any whitespace
+                int nameStart = keywordPos + keyword.length();
+                while (nameStart < sourceText.length() && Character.isWhitespace(sourceText.charAt(nameStart))) {
+                    nameStart++;
+                }
+
+                // Verify we've found the correct position by checking if the text matches the class name
+                String className = classDecl.name.toString();
+                if (sourceText.regionMatches(nameStart, className, 0, className.length())) {
+                    return nameStart;
+                }
+            }
+        }
+
+        return Position.NOPOS;
+    }
+
+
+    public Token toToken(int pos) {
+        if (pos == Position.NOPOS) {
+            return null;
+        }
+        return new Token(
+                compilationUnit.getLineMap().getLineNumber(pos),
+                compilationUnit.getLineMap().getColumnNumber(pos) + 1);
+    }
+}
